receipt_ocr-master/requirements.txt
flask
pillow
pytesseract
oauth2client
gspread
